refactor code

+++ component/site/views/product/tmpl/default.php
@@ -160,11 +160,11 @@ if (strstr($template_desc, '{navigation_link_right}') || strstr($template_desc, 
 			'&Itemid=' . $this->itemId
 		);
 
-		if (Redshop::getConfig()->get('DEFAULT_LINK_FIND') === 0)
+		if ((int) Redshop::getConfig()->get('DEFAULT_LINK_FIND') === 0)
 		{
 			$nextbutton = '<a href="' . $nextlink . '">' . $nextproducts->product_name . "" . Redshop::getConfig()->get('DAFULT_NEXT_LINK_SUFFIX') . '</a>';
 		}
-        elseif (Redshop::getConfig()->get('DEFAULT_LINK_FIND') == 1)
+        elseif ((int) Redshop::getConfig()->get('DEFAULT_LINK_FIND') == 1)
 		{
 			$nextbutton = '<a href="' . $nextlink . '">' . Redshop::getConfig()->get('CUSTOM_NEXT_LINK_FIND') . '</a>';
 		}
@@ -1573,7 +1573,7 @@ if (strstr($template_desc, "{product_rating_summary}"))
 
 if (strstr($template_desc, "{product_rating}"))
 {
-	if (Redshop::getConfig()->get('FAVOURED_REVIEWS') !== "" || Redshop::getConfig()->get('FAVOURED_REVIEWS') !== 0)
+	if ((int) Redshop::getConfig()->get('FAVOURED_REVIEWS') !== "" || (int) Redshop::getConfig()->get('FAVOURED_REVIEWS') !== 0)
 	{
 		$mainblock = Redshop::getConfig()->get('FAVOURED_REVIEWS');
 	}
